New problems to close Chapter 2
* Even numbers for photos! (1)
* Even numbers for photos! (2)

These introduce the idea of finding subsequences. In this case, it's
finding "chains of odd numbers" in a list of numbers. The first one just
identifies the chains. The second one finds the length and location of
the longest chain.

These have significant scaffolding with trace tables, and use new pages
to do so.

Added to 000-macros.tex:
 * \questionheadercontinued{...}, which allows us to (deliberately)
   continue a question over the page

Added to Informatics-Beginners.tex:
 * \robustify\bfseries, which allows us to make selected numbers in a
   table bold while maintaining the nice alignment. The following magic
   line is also required, which I put just before the table. It's
   probably a global thing, but nevermind (for now).
     \sisetup{detect-weight=true,detect-inline-weight=math}

Now that I know how to format numbers in tables more nicely, I'll apply
that to other tables in the next commit.